Reorganized into CLI and parser packages.
