+++ parser/src/tests.rs
@@ -1,5 +1,19 @@
+use crate::to_json;
+
+const BASIC_JSON: &str = "[{\"Subset\":{\"subsets\":[{\"Single\":{\"distance\":125,\"data\":{\"label\":null,\"reps\":5,\"attributes\":[\"75 swim\",\"@1:45\",\"25 drill\",\"25 kick\"]}}},{\"Single\":{\"distance\":200,\"data\":{\"label\":null,\"reps\":1,\"attributes\":[\"IM\",\"Drill / Swim\"]}}},{\"Single\":{\"distance\":50,\"data\":{\"label\":null,\"reps\":4,\"attributes\":[\"Choice\",\"@:55\",\"Variable\"]}}}],\"data\":{\"label\":\"Warm Up\",\"reps\":1,\"attributes\":[]}}}]";
 #[test]
 fn it_works() {
-    let result = 2 + 2;
-    assert_eq!(result, 4);
+    assert_eq!(1, 1)
 }
+
+#[test]
+fn basic() {
+    assert_eq!(to_json(r#"
+Warm Up: {
+    5 * 125 75 swim, @1:45 , 25 drill, 25 kick
+    200 IM, Drill / Swim
+    4 * 50 Choice, @:55, Variable
+}
+    "#).unwrap(), BASIC_JSON)
+}
+
