--- test.json
@@ -1,274 +1,0 @@
-[
-  {
-    "Text": "A test practice"
-  },
-  {
-    "Subset": {
-      "subsets": [
-        {
-          "Single": {
-            "distance": 125,
-            "data": {
-              "label": null,
-              "reps": 5,
-              "attributes": [
-                "75 swim",
-                "@1:45",
-                "25 drill",
-                "25 kick"
-              ]
-            }
-          }
-        },
-        {
-          "Single": {
-            "distance": 200,
-            "data": {
-              "label": null,
-              "reps": 1,
-              "attributes": [
-                "IM",
-                "Drill / Swim"
-              ]
-            }
-          }
-        },
-        {
-          "Single": {
-            "distance": 50,
-            "data": {
-              "label": null,
-              "reps": 4,
-              "attributes": [
-                "Choice",
-                "@:55",
-                "Variable"
-              ]
-            }
-          }
-        }
-      ],
-      "data": {
-        "label": "Warm Up",
-        "reps": 1,
-        "attributes": []
-      }
-    }
-  },
-  {
-    "Subset": {
-      "subsets": [
-        {
-          "Single": {
-            "distance": 25,
-            "data": {
-              "label": null,
-              "reps": 4,
-              "attributes": [
-                "IMO",
-                "@:45"
-              ]
-            }
-          }
-        },
-        {
-          "Single": {
-            "distance": 50,
-            "data": {
-              "label": null,
-              "reps": 6,
-              "attributes": [
-                "Free",
-                "@1:00",
-                "Maxi Mini"
-              ]
-            }
-          }
-        },
-        {
-          "Single": {
-            "distance": 125,
-            "data": {
-              "label": null,
-              "reps": 2,
-              "attributes": [
-                "Choice",
-                "@1:25/1:30/1:40",
-                "Just make it"
-              ]
-            }
-          }
-        }
-      ],
-      "data": {
-        "label": "Prep Set",
-        "reps": 1,
-        "attributes": []
-      }
-    }
-  },
-  {
-    "Subset": {
-      "subsets": [
-        {
-          "Subset": {
-            "subsets": [
-              {
-                "Single": {
-                  "distance": 100,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Free",
-                      "@1:20/1:30"
-                    ]
-                  }
-                }
-              },
-              {
-                "Single": {
-                  "distance": 200,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Free",
-                      "@2:40/3:00"
-                    ]
-                  }
-                }
-              },
-              {
-                "Single": {
-                  "distance": 300,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Free",
-                      "@4:00/4:30"
-                    ]
-                  }
-                }
-              },
-              {
-                "Single": {
-                  "distance": 400,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Free",
-                      "@5:20/6:00"
-                    ]
-                  }
-                }
-              },
-              {
-                "Text": "Very short rest to get pull stuff together"
-              },
-              {
-                "Single": {
-                  "distance": 400,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Pull",
-                      "@5:20/6:00"
-                    ]
-                  }
-                }
-              },
-              {
-                "Single": {
-                  "distance": 300,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Pull",
-                      "@4:00/4:30"
-                    ]
-                  }
-                }
-              },
-              {
-                "Single": {
-                  "distance": 200,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Pull",
-                      "@2:40/3:00"
-                    ]
-                  }
-                }
-              },
-              {
-                "Single": {
-                  "distance": 100,
-                  "data": {
-                    "label": null,
-                    "reps": 1,
-                    "attributes": [
-                      "Pull",
-                      "@1:20/1:30"
-                    ]
-                  }
-                }
-              }
-            ],
-            "data": {
-              "label": null,
-              "reps": 3,
-              "attributes": []
-            }
-          }
-        }
-      ],
-      "data": {
-        "label": "Main Set",
-        "reps": 1,
-        "attributes": []
-      }
-    }
-  },
-  {
-    "Subset": {
-      "subsets": [
-        {
-          "Single": {
-            "distance": 50,
-            "data": {
-              "label": null,
-              "reps": 1,
-              "attributes": [
-                "Destroy"
-              ]
-            }
-          }
-        },
-        {
-          "Single": {
-            "distance": 50,
-            "data": {
-              "label": null,
-              "reps": 1,
-              "attributes": [
-                "EZ"
-              ]
-            }
-          }
-        }
-      ],
-      "data": {
-        "label": "Cool Down",
-        "reps": 1,
-        "attributes": []
-      }
-    }
-  }
-]
