--- test.swim
@@ -1,31 +1,0 @@
-- A test practice
-Warm Up: {
-    5 * 125 75 swim, @1:45 , 25 drill, 25 kick
-    200 IM, Drill / Swim
-    4 * 50 Choice, @:55, Variable
-}
-
-Prep Set: {
-    4 * 25 IMO, @:45
-    6 * 50 Free, @1:00, Maxi Mini
-    2 * 125 Choice, @1:25/1:30/1:40, Just make it
-}
-
-Main Set: {
-    3 * {
-        100 Free, @1:20/1:30
-        200 Free, @2:40/3:00
-        300 Free, @4:00/4:30
-        400 Free, @5:20/6:00
-        - Very short rest to get pull stuff together
-        400 Pull, @5:20/6:00
-        300 Pull, @4:00/4:30
-        200 Pull, @2:40/3:00
-        100 Pull, @1:20/1:30
-    }
-}
-
-Cool Down: {
-    50 Destroy
-    50 EZ
-}
